🔧 chore(eslint, prettier, package.json): fix linting and formatting issues The changes in .eslintrc.cjs and .prettierrc fix the formatting issues in the codebase. The package.json file now includes new dependencies for the project.
✨ feat(package.json): add new dependencies
The new dependencies added to the project are:
- @fontsource/jetbrains-mono
- @fontsource/manrope
- lucide-svelte
- open-props

🔧 chore(app.d.ts): remove unnecessary semicolon
The semicolon at the end of the file was removed.

✨ feat(app.html): add new favicon and update head
The new favicon was added to the head of the HTML file. The head was also updated to include a new meta tag and the new favicon.

✨ feat(config.ts): add new config file
The new config file includes constants for the title, description, and